moved styles to a separate file
