+++ src/pages/Auth/Auth.tsx
@@ -7,63 +7,14 @@ import {
   IconButton,
   Box,
 } from "@material-ui/core";
-import { makeStyles } from "@material-ui/core";
 import VisibilityIcon from "@material-ui/icons/Visibility";
 import VisibilityOffIcon from "@material-ui/icons/VisibilityOff";
 import { Link } from "react-router-dom";
-
-const useStyles = makeStyles(() => ({
-  typography: {
-    marginBottom: "10px",
-  },
-  grid: {
-    alignItems: "center",
-    justifyContent: "center",
-    flexDirection: "column",
-    minHeight: "60vh",
-  },
-  box: {
-    padding: "50px",
-  },
-  textField: {
-    marginBottom: "20px",
-  },
-  emailInput: {
-    fontSize: "22px",
-  },
-  passwordInput: {
-    fontSize: "22px",
-    paddingRight: "44px",
-  },
-  passwordDiv: {
-    display: "flex",
-    position: "relative",
-  },
-  buttonsDiv: {
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-  },
-  link: {
-    fontSize: "24px",
-  },
-  button: {
-    fontSize: "16px",
-    paddingRight: "30px",
-    paddingLeft: "30px",
-    textTransform: "none",
-  },
-  icon: {
-    position: "absolute",
-    right: 0,
-    top: "10px",
-    padding: "10px",
-  },
-}));
+import useAuthStyles from "./theme";
 
 const Auth: React.FC = () => {
   const [visiblePassword, setVisiblePassword] = useState(false);
-  const classes = useStyles();
+  const classes = useAuthStyles();
 
   return (
     <form action="">
