+++ src/pages/Expenses.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-const Expenses = () => {
+const Expenses: React.FC = () => {
   return <div>expenses page</div>;
 };
